refactor: rename infuraService to wsProviderService
